# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PortalHatal Balmas API",
        "version": "2"
    },
    "paths": {
        "/api/announcements": {
            "get": {
                "tags": [
                    "Announcements"
                ],
                "summary": "מביא רשימת הכרזות",
                "description": "מביא רשימת הכרזות.",
                "operationId": "getAnnouncements",
                "responses": {
                    "200": {
                        "description": "הפעולה בוצעה בהצלחה"
                    },
                    "500": {
                        "description": "שגיאה בשרת"
                    }
                }
            },
            "post": {
                "tags": [
                    "Announcements"
                ],
                "summary": "יוצר הכרזה חדשה",
                "description": "יוצר הכרזה חדש עם כותרת, תיאור, תוכן,  ותמונה.",
                "operationId": "createAnnouncements",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "content",
                                    "image"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "כותרת ההכרזה"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "תיאור קצר של ההכרזה"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "{blah:{dfgfdgfd}"
                                    },
                                    "image": {
                                        "description": "תמונה מצורפת להכרזה",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ההכרזה נוצר בהצלחה"
                    },
                    "400": {
                        "description": "בקשה לא תקינה"
                    },
                    "500": {
                        "description": "שגיאה בשרת"
                    }
                }
            }
        },
        "/api/announcements/admin": {
            "get": {
                "tags": [
                    "Announcements"
                ],
                "summary": "מביא רשימת הכרזות",
                "description": "מביא רשימת הכרזות.",
                "operationId": "getAdminAnnouncements",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "חיפוש הכרזות",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": ""
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "הפעולה בוצעה בהצלחה"
                    },
                    "500": {
                        "description": "שגיאה בשרת"
                    }
                }
            }
        },
        "/api/announcements/{uuid}": {
            "get": {
                "tags": [
                    "Announcements"
                ],
                "summary": "מביא פוסט לפי UUID",
                "description": "מביא פוסט בודד לפי מזהה ייחודי (UUID).",
                "operationId": "getAnnouncementByUUid",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID של הכרזה",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "4a206b4-99d6-4692-915c-4935766e0420"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "הפעולה בוצעה בהצלחה"
                    },
                    "400": {
                        "description": "בקשה לא תקינה"
                    },
                    "404": {
                        "description": "הכרזה לא נמצא"
                    },
                    "500": {
                        "description": "שגיאה בשרת"
                    }
                }
            },
            "post": {
                "tags": [
                    "Announcements"
                ],
                "summary": "מעדכן הכזרזות ",
                "description": "מעדכן הכרזות .",
                "operationId": "updateAnnouncement",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "מערך של הכרזות עם שינויים",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "457946b0-6c14-4102-bf4a-675c61b228d1"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "כותרת ההכרזה"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "תיאור קצר של ההכרזה"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "{blah:{dfgfdgfd}"
                                    },
                                    "position": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "isVisible": {
                                        "description": "בוליאני מקבל רק 0 או 1",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "image": {
                                        "description": "תמונה מצורפת להכרזה",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "הכרזה עודכנה בהצלחה"
                    },
                    "400": {
                        "description": "בקשה לא תקינה"
                    },
                    "500": {
                        "description": "שגיאה בשרת"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Announcements"
                ],
                "summary": "מוחק הכרזה לפי UUID",
                "description": "מוחק הכרזה בודד לפי מזהה ייחודי (UUID).",
                "operationId": "deleteaAnnouncementsByUUid",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID של ההכרזה",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "4a206b4-99d6-4692-915c-4935766e0420"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "הפעולה בוצעה בהצלחה"
                    },
                    "404": {
                        "description": "הכרזה לא נמצא"
                    },
                    "500": {
                        "description": "שגיאה בשרת"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Announcements"
                ],
                "summary": "עדכון נראות של הכרזה",
                "description": "עדכון מאפיין ה- isVisible של הכרזה מסוימת לפי UUID.",
                "operationId": "updateAnnouncementVisibility",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID של ההכרזה",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "457946b0-6c14-4102-bf4a-675c61b228d1"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "isVisible"
                                ],
                                "properties": {
                                    "isVisible": {
                                        "description": "קובע האם ההכרזה גלויה או מוסתרת",
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "נראות ההכרזה עודכנה בהצלחה"
                    },
                    "404": {
                        "description": "הכרזה לא נמצאה"
                    },
                    "500": {
                        "description": "שגיאה בשרת"
                    }
                }
            }
        },
        "/api/announcements/updatePosition": {
            "put": {
                "tags": [
                    "Announcements"
                ],
                "summary": "עדכון מיקומים של מספר הכרזות",
                "description": "מעודכן את המיקומים של מספר הכרזות בהתבסס על הנתונים שנשלחו.",
                "operationId": "updateAnnouncementPosition",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "items": {
                                    "properties": {
                                        "uuid": {
                                            "type": "string",
                                            "format": "uuid",
                                            "example": "4a206b4-99d6-4692-915c-4935766e0420"
                                        },
                                        "position": {
                                            "type": "integer",
                                            "example": 2
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "המיקומים עודכנו בהצלחה",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "message": "הפעולה בוצעה בהצלחה"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "בקשה לא תקינה"
                    },
                    "404": {
                        "description": "אחת או יותר מההכרזות לא נמצאה"
                    },
                    "500": {
                        "description": "שגיאת שרת פנימית"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}